+++ lib/utils/app_helper.dart
@@ -13,5 +13,4 @@ class AppHelper {
           );
         });
   }
-
 }